Move sequential debuggers to execution framework
