+++ framework/execution_framework/plugins/org.eclipse.gemoc.executionframework.debugger.ui/META-INF/MANIFEST.MF
@@ -6,7 +6,8 @@ Bundle-Version: 3.0.0.qualifier
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-Activator: org.eclipse.gemoc.executionframework.debugger.ui.Activator
 Export-Package: org.eclipse.gemoc.executionframework.debugger.ui;uses:="org.eclipse.gemoc.xdsmlframework.api.core,org.eclipse.core.commands,org.eclipse.jface.viewers",
- org.eclipse.gemoc.executionframework.debugger.ui.breakpoints
+ org.eclipse.gemoc.executionframework.debugger.ui.breakpoints,
+ org.eclipse.gemoc.executionframework.debugger.ui.handlers
 Require-Bundle: com.google.guava,
  org.eclipse.xtext.xbase.lib,
  org.eclipse.xtend.lib,
