+++ framework/execution_framework/plugins/org.eclipse.gemoc.executionframework.debugger/META-INF/MANIFEST.MF
@@ -25,7 +25,9 @@ Require-Bundle: com.google.guava,
  org.eclipse.xtext;bundle-version="2.10.0",
  org.eclipse.gemoc.commons.eclipse;bundle-version="2.4.0",
  org.eclipse.gemoc.commons.eclipse.pde;bundle-version="2.4.0",
- org.eclipse.gemoc.commons.eclipse.messagingsystem.api;bundle-version="2.4.0"
-Export-Package: org.eclipse.gemoc.executionframework.debugger
+ org.eclipse.gemoc.commons.eclipse.messagingsystem.api;bundle-version="2.4.0",
+ fr.inria.diverse.melange.resource;bundle-version="0.2.1",
+ org.eclipse.gemoc.trace.gemoc.api;bundle-version="3.0.0"
 Bundle-ActivationPolicy: lazy
+Export-Package: org.eclipse.gemoc.executionframework.debugger
