+++ framework/execution_framework/plugins/org.eclipse.gemoc.executionframework.debugger/src/org/eclipse/gemoc/executionframework/debugger/GenericSequentialModelDebugger.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     Inria - initial API and implementation
  *******************************************************************************/
-package org.eclipse.gemoc.execution.sequential.javaengine.ui.debug;
+package org.eclipse.gemoc.executionframework.debugger;
 
 import java.util.ArrayDeque;
 import java.util.ArrayList;
@@ -21,9 +21,6 @@ import org.eclipse.emf.ecore.EObject;
 import org.eclipse.emf.ecore.EcorePackage;
 import org.eclipse.emf.ecore.resource.Resource;
 import org.eclipse.gemoc.dsl.debug.ide.event.IDSLDebugEventProcessor;
-import org.eclipse.gemoc.execution.sequential.javaengine.PlainK3ExecutionEngine;
-import org.eclipse.gemoc.executionframework.debugger.AbstractGemocDebugger;
-import org.eclipse.gemoc.executionframework.debugger.GemocBreakpoint;
 import org.eclipse.gemoc.executionframework.engine.core.EngineStoppedException;
 import org.eclipse.gemoc.trace.commons.model.trace.MSE;
 import org.eclipse.gemoc.trace.commons.model.trace.MSEOccurrence;
@@ -33,6 +30,7 @@ import org.eclipse.xtext.naming.DefaultDeclarativeQualifiedNameProvider;
 import org.eclipse.xtext.naming.QualifiedName;
 
 import fr.inria.diverse.melange.resource.MelangeResourceImpl;
+
 
 public class GenericSequentialModelDebugger extends AbstractGemocDebugger {
 
@@ -232,7 +230,6 @@ public class GenericSequentialModelDebugger extends AbstractGemocDebugger {
 		EObject target = o;
 		// Try to get the original object if 'o' comes from 
 		// a downcast resource
-		if(this.engine instanceof PlainK3ExecutionEngine){
 			Resource res = o.eResource();
 			if(res != null) {
 				
@@ -249,7 +246,6 @@ public class GenericSequentialModelDebugger extends AbstractGemocDebugger {
 					target = mr.getWrappedResource().getEObject(uriFragment);
 				}
 			}
-		}
 		
 		return super.shouldBreak(target)
 				&& (Boolean.valueOf((String) getBreakpointAttributes(target, GemocBreakpoint.BREAK_ON_LOGICAL_STEP)) || Boolean
