+++ java_execution/java_engine/plugins/org.eclipse.gemoc.execution.sequential.javaengine.ui/src/org/eclipse/gemoc/execution/sequential/javaengine/ui/debug/GemocSequentialDebugAs.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     Inria - initial API and implementation
  *******************************************************************************/
-package org.eclipse.gemoc.execution.sequential.javaengine.ui.debug.sirius.action;
+package org.eclipse.gemoc.execution.sequential.javaengine.ui.debug;
 
 
 import org.eclipse.gemoc.execution.sequential.javaengine.ui.launcher.Launcher;
